feat(datasets): spans as examples (#3279)
* feat(datasets): spans as examples

* WIP

* feat(datasets): span as an example resolver

* Update src/phoenix/server/api/types/Span.py


* refactor test

* Update src/phoenix/server/api/types/Span.py


* Add Example interface

* Revert frozen

* move tests to their own file

* move tests to their own file

---------
