+++ app/schema.graphql
@@ -103,7 +103,7 @@ type DatasetEdge {
   node: Dataset!
 }
 
-type DatasetExample implements Node {
+type DatasetExample implements Node & Example {
   """The Globally Unique ID of this object"""
   id: GlobalID!
   input: JSON!
@@ -495,6 +495,12 @@ type EventMetadata {
   actualLabel: String
 }
 
+interface Example {
+  input: JSON!
+  output: JSON!
+  metadata: JSON!
+}
+
 type ExportedFile {
   """File name without the file extension."""
   fileName: String!
@@ -866,6 +872,15 @@ type Span implements Node {
 
   """All descendant spans (children, grandchildren, etc.)"""
   descendants: [Span!]!
+
+  """The span's attributes translated into an example for a dataset"""
+  asExample: SpanAsExample!
+}
+
+type SpanAsExample implements Example {
+  input: JSON!
+  output: JSON!
+  metadata: JSON!
 }
 
 enum SpanColumn {
