+++ src/phoenix/server/api/types/Span.py
@@ -1,4 +1,5 @@
 import json
+from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
 from typing import Any, List, Mapping, Optional, Sized, cast
@@ -13,8 +14,13 @@ from strawberry.types import Info
 import phoenix.trace.schemas as trace_schema
 from phoenix.db import models
 from phoenix.server.api.context import Context
+from phoenix.server.api.helpers.dataset_helpers import (
+    get_dataset_example_input,
+    get_dataset_example_output,
+)
 from phoenix.server.api.types.DocumentRetrievalMetrics import DocumentRetrievalMetrics
 from phoenix.server.api.types.Evaluation import DocumentEvaluation, SpanEvaluation
+from phoenix.server.api.types.ExampleInterface import Example
 from phoenix.server.api.types.MimeType import MimeType
 from phoenix.trace.attributes import get_attribute_value
 
@@ -25,6 +31,9 @@ INPUT_VALUE = SpanAttributes.INPUT_VALUE
 LLM_TOKEN_COUNT_COMPLETION = SpanAttributes.LLM_TOKEN_COUNT_COMPLETION
 LLM_TOKEN_COUNT_PROMPT = SpanAttributes.LLM_TOKEN_COUNT_PROMPT
 LLM_TOKEN_COUNT_TOTAL = SpanAttributes.LLM_TOKEN_COUNT_TOTAL
+LLM_PROMPT_TEMPLATE_VARIABLES = SpanAttributes.LLM_PROMPT_TEMPLATE_VARIABLES
+LLM_INPUT_MESSAGES = SpanAttributes.LLM_INPUT_MESSAGES
+LLM_OUTPUT_MESSAGES = SpanAttributes.LLM_OUTPUT_MESSAGES
 METADATA = SpanAttributes.METADATA
 OUTPUT_MIME_TYPE = SpanAttributes.OUTPUT_MIME_TYPE
 OUTPUT_VALUE = SpanAttributes.OUTPUT_VALUE
@@ -102,8 +111,13 @@ class SpanEvent:
 
 
 @strawberry.type
+class SpanAsExample(Example): ...
+
+
+@strawberry.type
 class Span(Node):
     id_attr: NodeID[int]
+    db_span: strawberry.Private[models.Span]
     name: str
     status_code: SpanStatusCode
     status_message: str
@@ -189,6 +203,31 @@ class Span(Node):
         spans = await info.context.data_loaders.span_descendants.load(span_id)
         return [to_gql_span(span) for span in spans]
 
+    @strawberry.field(
+        description="The span's attributes translated into an example for a dataset",
+    )  # type: ignore
+    def as_example(self) -> SpanAsExample:
+        db_span = self.db_span
+        attributes = db_span.attributes
+        span_io = _SpanIO(
+            span_kind=db_span.span_kind,
+            input_value=get_attribute_value(attributes, INPUT_VALUE),
+            input_mime_type=get_attribute_value(attributes, INPUT_MIME_TYPE),
+            output_value=get_attribute_value(attributes, OUTPUT_VALUE),
+            output_mime_type=get_attribute_value(attributes, OUTPUT_MIME_TYPE),
+            llm_prompt_template_variables=get_attribute_value(
+                attributes, LLM_PROMPT_TEMPLATE_VARIABLES
+            ),
+            llm_input_messages=get_attribute_value(attributes, LLM_INPUT_MESSAGES),
+            llm_output_messages=get_attribute_value(attributes, LLM_OUTPUT_MESSAGES),
+            retrieval_documents=get_attribute_value(attributes, RETRIEVAL_DOCUMENTS),
+        )
+        return SpanAsExample(
+            input=get_dataset_example_input(span_io),
+            output=get_dataset_example_output(span_io),
+            metadata=attributes,
+        )
+
 
 def to_gql_span(span: models.Span) -> Span:
     events: List[SpanEvent] = list(map(SpanEvent.from_dict, span.events))
@@ -198,6 +237,7 @@ def to_gql_span(span: models.Span) -> Span:
     num_documents = len(retrieval_documents) if isinstance(retrieval_documents, Sized) else None
     return Span(
         id_attr=span.id,
+        db_span=span,
         name=span.name,
         status_code=SpanStatusCode(span.status_code),
         status_message=span.status_message,
@@ -303,3 +343,21 @@ def _convert_metadata_to_string(metadata: Any) -> Optional[str]:
         return json.dumps(metadata)
     except Exception:
         return str(metadata)
+
+
+@dataclass
+class _SpanIO:
+    """
+    An class that contains the information needed to extract dataset example
+    input and output values from a span.
+    """
+
+    span_kind: Optional[str]
+    input_value: Any
+    input_mime_type: Optional[str]
+    output_value: Any
+    output_mime_type: Optional[str]
+    llm_prompt_template_variables: Any
+    llm_input_messages: Any
+    llm_output_messages: Any
+    retrieval_documents: Any
